Add files via upload
Problem Statement
Roger Ferro, the Country head of a multi-national retail company is responsible for understanding the sales numbers and to make them grow. For this, he wants to gauge and execute suitable sales techniques, to scale up the sales numbers for various products, retailed by various vendors.
The country heads wants to analyze the stock, price, number of items sold and quality rating for various categories to find the sales trends. He would also want to evaluate the inventory status for various categories, subcategories, and items.
With such meticulous exploration of vendor sales and inventory stocks, the country head can formulate schemes to improvise sales in low profit periods and confirm that the inventory has enough stock to meet the customer requirements.